refactor: refactor with webpack
