+++ src/config.js
@@ -1,0 +1,14 @@
+export class config {}
+config.checkDomain = "jwstu.ustb.edu.cn";
+config.jsonUrl = "http://jwstu.ustb.edu.cn/smvc/StuQueryInfoService/viewStuCourseSchedule.json";
+config.initialDay = "2020-02-24";
+config.alarmOn = true;
+config.classTable = [
+    [0, 0, 0, 0],
+    [8, 0, 9, 35],
+    [9, 55, 11, 30],
+    [13, 30, 15, 5],
+    [15, 20, 16, 55],
+    [17, 10, 18, 45],
+    [19, 30, 21, 5]
+];
