+++ src/instance.js
@@ -1,0 +1,4 @@
+import icsFormatter from "./ics-formatter";
+import {config} from "./config";
+
+export let icsObj = new icsFormatter(config.alarmOn);
